requirements.txt
pandas
yfinance
rapidfuzz
requests
google-cloud-bigquery
db-dtypes